refactor Markdown into it’s own file
